0528final

+++ src/main/java/com/haiso/commons/enumeration/ExcelCellDataType.java
@@ -4,5 +4,5 @@ package com.haiso.commons.enumeration;
  * Created by Heli on 2014/5/17.
  */
 public enum ExcelCellDataType {
-    STRING, NUMERIC, DATE, FORMULA, BOOLEAN, BLANK, ERROR, EMPTY
+    STRING, DOUBLE, DATE, FORMULA, BOOLEAN, BLANK, ERROR, EMPTY
 }
